+++ packages/@cdk-cosmos/core/src/features/cicd-solar-system/cicd-solar-system-core-stack.ts
@@ -37,11 +37,11 @@ export class CiCdFeatureCoreStack extends BaseFeatureStack implements ICiCdFeatu
         mutable: false,
       }),
       deployStacks: [this.solarSystem.nodeId('*', '', CDK_PIPELINE_STACK_PATTERN)],
+      deploySubnets: { subnetGroupName: 'App' },
       ...cdkPipelineProps,
       pipelineName: this.solarSystem.nodeId('Cdk-Pipeline', '-', CDK_PIPELINE_PATTERN),
       deployName: this.solarSystem.nodeId('Cdk-Deploy', '-', CDK_PIPELINE_PATTERN),
       cdkRepo: cdkRepo,
-      // deployVpc: vpc,
     });
     this.deployProject = this.deployPipeline.deploy;
   }